Add captive portal detection
This patch detects whether a captive portal is active and adds a simple
button to open it in the default web browser to allow easy login to
public wifi APs.

The DBus messaging has also been reworked to avoid passing an excessive
number of functions to the `wifi_listen` function.